gather not working yet

+++ src/distributed/smesh_distributed_read.cpp
@@ -29,7 +29,7 @@ template int mesh_from_folder<i32, f32, i32>(
     ptrdiff_t *nelements_out, i32 ***elements_out, int *spatial_dim_out,
     ptrdiff_t *nnodes_out, f32 ***points_out, ptrdiff_t *n_owned_nodes_out,
     ptrdiff_t *n_owned_elements_out, i32 **element_mapping_out,
-    i32 **node_mapping_out, int **node_owner_out, i32 **node_offsets_out,
+    i32 **node_mapping_out, int **node_owner_out, ptrdiff_t **node_offsets_out,
     i32 **ghosts_out, ptrdiff_t *n_owned_nodes_with_ghosts_out,
     ptrdiff_t *n_shared_elements_out,
     ptrdiff_t *n_owned_elements_with_ghosts_out);
